+++ stylo/image/image.py
@@ -1,7 +1,9 @@
-from abc import ABC, abstractmethod
-
+import base64
+import io
 import PIL as P
 import matplotlib.pyplot as plt
+
+from abc import ABC, abstractmethod
 
 
 class Drawable:
@@ -15,33 +17,81 @@ class Drawable:
 
 
 class Image(ABC):
-    def __call__(self, width, height, filename=None, plot_size=None):
+    def __call__(
+        self, width, height, filename=None, plot_size=None, encode=None, preview=True
+    ):
 
-        image = self._render(width, height)
-        self.data = image
+        self.data = self._render(width, height)
 
-        if filename is not None:
-            self._save(image, filename)
+        if filename:
+            self.save(filename)
             return
+
+        if encode:
+            return self.encode()
+
+        if preview:
+            return self.preview(plot_size)
+
+    def preview(self, plot_size):
+        """Generate a matplotlib plot of the image data which can then be viewed from
+        a Jupyter Notebook.
+
+        :param plot_size: This controls how large the generated plot is
+        :type plot_size: int
+
+        :returns: A matplotlib AxesImage object.
+        """
 
         if plot_size is None:
             plot_size = 12
 
         fig, ax = plt.subplots(1, figsize=(plot_size, plot_size))
-        ax.imshow(image)
 
-        return fig
+        # Hide the axis - show just the image
+        fig.axes[0].get_yaxis().set_visible(False)
+        fig.axes[0].get_xaxis().set_visible(False)
 
-    def _save(self, image, filename):
+        # Draw the image
+        ax.imshow(self.data)
 
-        height, width, _ = image.shape
+        # Return just the axis, jupyter notebooks will capture the figure and
+        # display it anyway.
+        return ax
 
-        pil_image = P.Image.frombuffer(
-            "RGB", (width, height), image, "raw", "RGB", 0, 1
-        )
+    def encode(self):
+        """Encode the image as a PNG represented by a base64 string.
+
+        :return: The image encoded as a base64 string.
+        """
+
+        image = self._to_pil_image()
+
+        with io.BytesIO() as byte_stream:
+            image.save(byte_stream, "PNG")
+            image_bytes = byte_stream.getvalue()
+
+        return base64.b64encode(image_bytes)
+
+    def save(self, filename):
+        """Save the image to file as a PNG image.
+
+        If the given filename already exists the existing image will be overwritten.
+
+        :param filename: The file to save the image to.
+        :type filename: str
+        """
+
+        image = self._to_pil_image()
 
         with open(filename, "wb") as f:
-            pil_image.save(f)
+            image.save(f)
+
+    def _to_pil_image(self):
+        """Convert the numpy representation of the image into a PIL.Image object."""
+
+        height, width, _ = self.data.shape
+        return P.Image.frombuffer("RGB", (width, height), self.data, "raw", "RGB", 0, 1)
 
     @property
     def data(self):